KVM: PPC: Book3S: Move vcore definition to end of kvm_arch struct
When building KVM with a lot of vcores (NR_CPUS is big), we can potentially
get out of the ld immediate range for dereferences inside that struct.

Move the array to the end of our kvm_arch struct. This fixes compilation
issues with NR_CPUS=2048 for me.

Signed-off-by: Alexander Graf <agraf@suse.de>

Gbp-Pq: Topic features/powerpc
Gbp-Pq: Name KVM-PPC-Book3S-Move-vcore-definition-to-end-of-kvm_a.patch