scripts: Fix X.509 PEM support in sign-file
sign-file originally required the X.509 certificate to be in DER
format, but now has a fallback to PEM format.  It expects BIO_reset()
to return 1 on success, but:

       BIO_reset() normally returns 1 for success and 0 or -1 for failure.
       File BIOs are an exception, they return 0 for success and -1 for
       failure.

BIO_reset() also prints accumulated error messages, which we don't
want when we're about to try a fallback, so drain them first.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name scripts-fix-x.509-pem-support-in-sign-file.patch